+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.2.1] (2026-05-24)
 ### Added
 * `make list` into `Makefile`
 * formalized Inlining Strategy documentation in `src/lib.rs` to guide future development
@@ -20,7 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * `clippy::useless_concat`
-
 
 ## [0.2.0] (2025-09-24)
 ### Added
@@ -205,7 +207,8 @@ bug: case of the needle is empty on `SearchIndices`
 ## [0.1.0] (2021-03-17)
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/naive_opt/compare/v0.2.0..HEAD
+[Unreleased]: https://github.com/aki-akaguma/naive_opt/compare/v0.2.1..HEAD
+[0.2.1]: https://github.com/aki-akaguma/naive_opt/compare/v0.2.0..v0.2.1
 [0.2.0]: https://github.com/aki-akaguma/naive_opt/compare/v0.1.25..v0.2.0
 [0.1.25]: https://github.com/aki-akaguma/naive_opt/compare/v0.1.24..v0.1.25
 [0.1.24]: https://github.com/aki-akaguma/naive_opt/compare/v0.1.23..v0.1.24